bugfix: Single day with 10000 results should not split date (#54)
If a search is for a single day then it does not matter if the request
returns more than 10,000 hits (the max the SEC returns in a single
search query). We cannot search with finer granularity than a single
day.

I also took the liberty of adding a formatter for Markdown files.